LayoutTests/platform/mac-wk1/fast/scrolling/latching/nested-cross-axis-latch-expiration-expected.txt

PASS innerOverflowScrollEventCount > 0 is true
FAIL outerOverflowScrollEventCount should be 0. Was 11.
PASS windowScrollEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
